Redesign the code

+++ src/kvraft/common.go
@@ -15,7 +15,7 @@ type PutAppendArgs struct {
 	// You'll have to add definitions here.
 	// Field names must start with capital letters,
 	// otherwise RPC will break.
-	Cid   		int
+	Cid   		int64
 	RequestId	int
 }
 
@@ -27,7 +27,7 @@ type PutAppendReply struct {
 type GetArgs struct {
 	Key string
 	// You'll have to add definitions here.
-	Cid int
+	Cid int64
 	RequestId int
 }
